requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.2
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.5
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
python-multipart==0.0.12
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
